backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
django-filter==21.1
django-cors-headers==3.7.0
drf-extra-fields==3.2.1
Pillow==9.0.0
python-dotenv==0.19.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
django-ratelimit==4.1.0

# Dev dependencies
flake8==4.0.1
isort==5.10.1
pytest==7.1.1
pytest-django==4.5.2
pytest-pythonpath==0.7.4

# Documentation
drf-yasg==1.20.0